requirements.txt
requests==2.31.0
pyserial==3.5
zeroconf==0.38.4
pydantic==2.6.3
python-dotenv==1.0.1